server/requirements.txt
gunicorn==20.1.0
Flask==2.0.2
Flask-Cors==3.0.10
Flask-CAS==1.0.2
flask_sqlalchemy==3.0.2
python-dateutil==2.8.2
six<2.0.0,>=1.13.0
jsonpickle==1.4.2